requirements.txt
langgraph==0.2.0
crewai==0.45.0
langchain-openai==0.1.8
langchain==0.1.14
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.25.2

# For Ollama-based inference (Local)
ollama==0.1.25

# For HuggingFace vLLM inference
vllm==0.5.0
torch==2.3.0
accelerate==0.27.0
huggingface-hub==0.21.0

# Optional: For GPU support
# nvidia-cuda-runtime-cu12==12.1.105
# cupy-cuda12x==13.0.0
